+++ apps/swirl-docs/src/components/Layout/DocLinksNav.tsx
@@ -1,8 +1,8 @@
-import { FunctionComponent, useEffect, useRef, useState } from "react";
+import { FunctionComponent, useEffect, useState } from "react";
 
 import { DocHeadline } from "@swirl/lib/docs/src/docs.model";
-import Link from "next/link";
 import useScrollObserver from "@swirl/lib/hooks/useScrollObserver";
+import Link from "next/link";
 
 type DocLinksNavProps = {
   documentLinkList: DocHeadline[];
@@ -13,13 +13,18 @@ const isBrowser = typeof window !== "undefined";
 export const DocLinksNav: FunctionComponent<DocLinksNavProps> = ({
   documentLinkList,
 }) => {
-  let headlines: Element[] = [];
+  let sections: Element[] = [];
 
   if (isBrowser) {
-    headlines = documentLinkList.flatMap((item) => {
-      const sectionElement = document.querySelector(`[id="${item.id}"]`);
+    sections = documentLinkList.flatMap((item) => {
+      const sectionElement = document
+        .querySelector(`[id="${item.id}"]`)
+        ?.closest<HTMLElement>("section");
+
       const subsectionElements = item.children.map((sub) =>
-        document.querySelector(`[id="${sub.id}"]`)
+        document
+          .querySelector(`[id="${sub.id}"]`)
+          ?.closest<HTMLElement>("section")
       );
 
       return [sectionElement, ...subsectionElements].filter(
@@ -29,7 +34,7 @@ export const DocLinksNav: FunctionComponent<DocLinksNavProps> = ({
   }
 
   const [activeIndex, setActiveIndexInLinks] = useState(0);
-  const [currentActiveIndex] = useScrollObserver(headlines);
+  const [currentActiveIndex] = useScrollObserver(sections);
 
   useEffect(() => {
     setActiveIndexInLinks(currentActiveIndex);
@@ -38,7 +43,7 @@ export const DocLinksNav: FunctionComponent<DocLinksNavProps> = ({
   return (
     <nav
       aria-label="document"
-      className="sticky top-[128px] h-max hidden min-w-[12rem] md:block px-4 border-l-1"
+      className="sticky top-[4rem] h-max hidden min-w-[12rem] md:block px-4 border-l-1"
     >
       <div className="mb-4 font-sm font-semibold text-text-subdued text-sm">
         On this Page
@@ -47,28 +52,27 @@ export const DocLinksNav: FunctionComponent<DocLinksNavProps> = ({
         {documentLinkList?.map((link: DocHeadline, index: number) => {
           return (
             <li key={link.id} className="relative font-sm mb-2">
-              <Link href={`#${link.id}`}>
-                <a
-                  className={`
-                    transition-colors duration-500 ease-in-out
-                    before:transition-colors before:duration-500 before:ease-in-out
-                    before:block before:absolute before:top-0 before:left-[-17px] before:w-[2px] before:h-6 text-sm
-                    ${
-                      activeIndex === index
-                        ? "text-border-info before:bg-border-info"
-                        : "text-text-subdued"
-                    }
-                  `}
-                  onClick={() => {
+              <Link
+                href={`#${link.id}`}
+                className={`
+                  transition-colors duration-500 ease-in-out
+                  before:transition-colors before:duration-500 before:ease-in-out
+                  before:block before:absolute before:top-0 before:left-[-17px] before:w-[2px] before:h-6 text-sm
+                  ${
+                    activeIndex === index
+                      ? "text-border-info before:bg-border-info"
+                      : "text-text-subdued"
+                  }
+                `}
+                onClick={() => {
+                  setActiveIndexInLinks(index);
+                  // Wait for the scroll observer to update the active index, then set the active index manually
+                  setTimeout(() => {
                     setActiveIndexInLinks(index);
-                    // Wait for the scroll observer to update the active index, then set the active index manually
-                    setTimeout(() => {
-                      setActiveIndexInLinks(index);
-                    }, 100);
-                  }}
-                >
-                  {link.title}
-                </a>
+                  }, 100);
+                }}
+              >
+                {link.title}
               </Link>
             </li>
           );