+++ apps/swirl-docs/src/lib/hooks/useScrollObserver.ts
@@ -13,11 +13,23 @@ const useScrollObserver = (
     }
 
     observer.current = new IntersectionObserver((entries) => {
-      const newHightlightIndex = entries.findIndex((entry) => {
+      const intersecting = entries.filter((entry) => {
         return entry.intersectionRatio > 0;
       });
 
-      setActiveIndex(newHightlightIndex);
+      const firstIntersectingElementHasIntersectingChild =
+        intersecting.length > 1 &&
+        intersecting[0].target.contains(intersecting[1].target);
+
+      if (firstIntersectingElementHasIntersectingChild) {
+        const index = entries.indexOf(intersecting[1]);
+        setActiveIndex(index);
+        return;
+      }
+
+      const index = entries.indexOf(intersecting[0]);
+
+      setActiveIndex(index);
     });
 
     const { current: currentObserver } = observer;