+++ apps/swirl-docs/CHANGELOG.md
@@ -1,5 +1,33 @@
 # swirl-docs
 
+## 0.1.235
+
+### Patch Changes
+
+- Updated dependencies
+  [[`d4567b59`](https://github.com/getflip/swirl/commit/d4567b592ab80a65545b57ce351a812de7b01f16)]:
+  - @getflip/swirl-components@0.118.3
+  - @getflip/swirl-components-react@0.118.3
+
+## 0.1.234
+
+### Patch Changes
+
+- Updated dependencies
+  [[`6c0affef`](https://github.com/getflip/swirl/commit/6c0affef835427fd76852dfa16b86d0b85a673a6)]:
+  - @getflip/swirl-components@0.118.2
+  - @getflip/swirl-components-react@0.118.2
+
+## 0.1.233
+
+### Patch Changes
+
+- Updated dependencies
+  [[`21bedf5b`](https://github.com/getflip/swirl/commit/21bedf5bc1a2e9c0a9cfca75da54a5a2080ed62f)]:
+  - @getflip/swirl-tokens@2.5.0
+  - @getflip/swirl-components@0.118.1
+  - @getflip/swirl-components-react@0.118.1
+
 ## 0.1.232
 
 ### Patch Changes