+++ packages/bridge/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @getflip/bridge
 
+## 0.3.6
+
+### Patch Changes
+
+- Updated dependencies
+  [[`d4567b59`](https://github.com/getflip/swirl/commit/d4567b592ab80a65545b57ce351a812de7b01f16)]:
+  - @getflip/swirl-components@0.118.3
+
+## 0.3.5
+
+### Patch Changes
+
+- Updated dependencies
+  [[`6c0affef`](https://github.com/getflip/swirl/commit/6c0affef835427fd76852dfa16b86d0b85a673a6)]:
+  - @getflip/swirl-components@0.118.2
+
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @getflip/swirl-components@0.118.1
+
 ## 0.3.3
 
 ### Patch Changes